No UI test are implemented. Final Clean up for PR.
