--- test/unit_test.dart
@@ -1,15 +1,0 @@
-// This is an example unit test.
-//
-// A unit test tests a single function, method, or class. To learn more about
-// writing unit tests, visit
-// https://flutter.dev/to/unit-testing
-
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  group('Plus Operator', () {
-    test('should add two numbers together', () {
-      expect(1 + 1, 2);
-    });
-  });
-}
